+++ packages/@lazy-random/df-char-id/src/index.ts
@@ -35,7 +35,7 @@ export function dfCharID(random: IRNGLike, char?: ENUM_ALPHABET | string | Buffe
 	const ls = UString.create(char).split('');
 	const len = ls.length;
 
-	expect(ls).to.have.lengthOf.gt(1);
+	expect(ls).lengthOf.gt(1);
 
 	const randIndex = () =>
 	{