+++ packages/@lazy-random/expect/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.7](https://github.com/bluelovers/ws-random/compare/@lazy-random/expect@1.0.5...@lazy-random/expect@1.0.7) (2021-12-20)
+
+
+### 🐛　Bug Fixes
+
+* update `@lazy-random/expect` ([fdd126b](https://github.com/bluelovers/ws-random/commit/fdd126b684ca684228f00b25dd0cc8db064c97e1))
+
+
+### 🔖　Miscellaneous
+
+* . ([c9f90c3](https://github.com/bluelovers/ws-random/commit/c9f90c3c7bacda06796085537bc4e9b56e96759a))
+
+
+
+
+
 ## [1.0.6](https://github.com/bluelovers/ws-random/compare/@lazy-random/expect@1.0.5...@lazy-random/expect@1.0.6) (2021-12-20)